Django/myshop/requirements.txt
amqp==5.0.6
asgiref==3.4.1
autopep8==1.6.0
billiard==3.6.4.0
braintree==4.12.0
Brotli==1.0.9
celery==5.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==35.0.0
cssselect2==0.4.1
dj-database-url==0.5.0
Django==3.2.11
django-extensions==3.1.5
django-localflavor==3.1
django-parler==2.2.1
django-rosetta==0.9.7
easy-thumbnails==2.8
flower==1.0.0
fonttools==4.27.1
gunicorn==20.1.0
html5lib==1.1
humanize==3.12.0
idna==3.3
kombu==5.1.0
lxml==4.6.5
Pillow==9.0.0
polib==1.1.1
prometheus-client==0.12.0
prompt-toolkit==3.0.21
psycopg2==2.9.2
pycodestyle==2.8.0
pycparser==2.20
pydyf==0.1.2
pyOpenSSL==21.0.0
pyphen==0.11.0
python-dotenv==0.19.1
python-stdnum==1.17
pytz==2021.3
redis==3.5.3
reportlab==3.6.2
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
svglib==1.1.0
tinycss2==1.1.0
toml==0.10.2
tornado==6.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
weasyprint==53.3
webencodings==0.5.1
Werkzeug==2.0.2
whitenoise==5.3.0
zopfli==0.1.9
